Migrate CreateExercise TS logic into a custom hook
