+++ frontend/kettlepal-fe/src/Components/NewWorkouts/CreateExercise.tsx
@@ -1,11 +1,7 @@
-import React, { useCallback, useEffect, useState } from "react";
+import React from "react";
 
-import { HStack, Button, useDisclosure } from "@chakra-ui/react";
+import { HStack, Button } from "@chakra-ui/react";
 import AddComment from "./FormComponents.tsx/Generic/AddComment";
-import {
-  ExerciseTitles,
-  KettlbellWeightsKG,
-} from "../../Constants/ExercisesOptions";
 import ConfirmModal from "../ConfirmModal";
 import theme from "../../Constants/theme";
 import ExerciseTitle from "./FormComponents.tsx/Exercise/ExerciseTitle";
@@ -18,6 +14,7 @@ import ExerciseTimer from "./FormComponents.tsx/Exercise/ExerciseTimer";
 import TrackExercise from "./FormComponents.tsx/MidWorkoutTracking/TrackExercise";
 import { ExerciseContainer } from "./FormComponents.tsx/Exercise/ExerciseContainer";
 import { CreateWorkoutState } from "../../Hooks/useCreateWorkoutForm";
+import useCreateExerciseForm from "../../Hooks/useCreateExerciseForm";
 
 interface CreateExerciseProps {
   exercise: Omit<CreateWorkoutState["exercises"][number], "key">;
@@ -40,182 +37,46 @@ export default function CreateExercise({
   trackWorkout,
   mutatingWorkout,
 }: CreateExerciseProps) {
-  const SESSION_STORAGE_KEY = `completedSets-${exerciseIndex}`;
-  const EXERCISE_TIMER_KEY = `exerciseTimerIsActive-${exerciseIndex}`;
-
-  const [completedSets, setCompletedSets] = useState<number>(() => {
-    const sessionVal = sessionStorage.getItem(SESSION_STORAGE_KEY);
-    return sessionVal ? parseInt(sessionVal) : 0;
-  });
-  const [seeDetails, setSeeDetails] = useState<boolean>(false);
-  const [timerIsActive, setTimerIsActive] = useState(() => {
-    const fromStorage = sessionStorage.getItem(EXERCISE_TIMER_KEY);
-    return fromStorage ? true : false;
-  });
-  const [customTitle, setCustomTitle] = useState(
-    exercise.title !== "" && !ExerciseTitles.includes(exercise.title)
-  );
-  const [customWeight, setCustomWeight] = useState(
-    exercise.weight !== "" && !KettlbellWeightsKG.includes(exercise.weight)
-  );
-
-  const handleTimerIsActive = (newState: boolean) => {
-    setTimerIsActive(newState);
-    if (newState) {
-      sessionStorage.setItem(EXERCISE_TIMER_KEY, "true");
-    } else {
-      sessionStorage.removeItem(EXERCISE_TIMER_KEY);
-    }
-  };
-  const setExerciseComment = (newComment: string) => {
-    handleExercise("comment", newComment, exerciseIndex);
-  };
-
-  function onDeleteExercise(): void {
-    deleteExercise(exerciseIndex);
-    onCloseDeleteExercise();
-    setOffset(
-      parseInt(
-        `${
-          !!swipeDistance() && swipeDistance() > minSwipeDistance
-            ? swipeDistance()
-            : offset
-        }px`
-      )
-    );
-  }
-
-  const setTime = useCallback(
-    (elapsedSeconds: number) => {
-      handleExercise("elapsedSeconds", elapsedSeconds, exerciseIndex);
-    },
-    [exerciseIndex, handleExercise]
-  );
-
-  // Update exercise timer every 1s
-  useEffect(() => {
-    let interval: NodeJS.Timeout | undefined = undefined;
-    if (timerIsActive) {
-      interval = setInterval(() => {
-        setTime(exercise.elapsedSeconds + 1);
-      }, 1000);
-    } else if (!timerIsActive && exercise.elapsedSeconds !== 0) {
-      clearInterval(interval);
-    }
-    return () => clearInterval(interval);
-  }, [timerIsActive, exercise.elapsedSeconds, setTime]);
-
-  // ERROR VALIDATION
-  const titleIsInvalid = !exercise.title;
-  const weightIsInvalid =
-    !!exercise.weight === false && !!exercise.weightUnit === true;
-  const weightUnitIsInvalid =
-    !!exercise.weight === true && !!exercise.weightUnit === false;
-  const setsIsInvalid =
-    (!!exercise.reps === true || !!exercise.repsDisplay === true) &&
-    !!exercise.sets === false;
-  const repsIsInvalid =
-    (!!exercise.sets === true || !!exercise.repsDisplay === true) &&
-    !!exercise.reps === false;
-  const repsDisplayIsInvalid =
-    (!!exercise.reps === true || !!exercise.sets === true) &&
-    !!exercise.repsDisplay === false;
-  const timerIsInvalid = timerIsActive;
-
-  enum ExerciseErrors {
-    title = "Title is required.",
-    weight = "Weight is required when Weight Unit is provided.",
-    weightUnit = "Weight Unit is required when Weight is proivided.",
-    sets = "Sets are required when Reps or Rep Type are provided.",
-    reps = "Reps are required when Sets or Rep Type are provided.",
-    repsDisplay = "Rep Type is required when Sets or Reps are provided.",
-    timer = "Please stop the exercise timer before saving.",
-  }
-
-  const [numErrors, setNumErrors] = useState(0);
-  const errors: string[] = [];
-  if (titleIsInvalid) errors.push(ExerciseErrors.title);
-  if (weightIsInvalid) errors.push(ExerciseErrors.weight);
-  if (weightUnitIsInvalid) errors.push(ExerciseErrors.weightUnit);
-  if (setsIsInvalid) errors.push(ExerciseErrors.sets);
-  if (repsIsInvalid) errors.push(ExerciseErrors.reps);
-  if (repsDisplayIsInvalid) errors.push(ExerciseErrors.repsDisplay);
-  if (timerIsInvalid) errors.push(ExerciseErrors.timer);
-
-  if (numErrors !== errors.length) {
-    setNumErrors(errors.length);
-  }
-
-  useEffect(
-    () => setFormHasErrors(numErrors > 0),
-    [numErrors, setFormHasErrors]
-  );
-
-  // Keep track of completed sets in session storage
-  useEffect(() => {
-    sessionStorage.setItem(SESSION_STORAGE_KEY, completedSets.toString());
-  }, [completedSets, SESSION_STORAGE_KEY]);
-
-  function completedASet() {
-    setCompletedSets((prev) => prev + 1);
-    if (exercise.sets === "") {
-      handleExercise("sets", "1", exerciseIndex);
-    } else if (completedSets >= parseInt(exercise.sets)) {
-      handleExercise("sets", (completedSets + 1).toString(), exerciseIndex);
-    }
-  }
-  function removedASet() {
-    if (completedSets === 0) {
-      return;
-    }
-    setCompletedSets((prev) => prev - 1);
-  }
-
-  // Bump the number of sets if your go above the limt while tracking
-  useEffect(() => {
-    if (completedSets > parseInt(exercise.sets)) {
-      setCompletedSets(parseInt(exercise.sets));
-    }
-  }, [exercise.sets, completedSets, setCompletedSets]);
-
-  /** SWIPE LOGIC **/
-  const [offset, setOffset] = useState<number>(0);
-  const [touchStart, setTouchStart] = useState(null);
-  const [touchEnd, setTouchEnd] = useState(null);
-  const minSwipeDistance = 50;
-  const swipeDistance = () => {
-    if (!touchStart || !touchEnd) {
-      return 0;
-    }
-    return touchStart - touchEnd;
-  };
-  const onTouchStart = (e: any) => {
-    setTouchEnd(null);
-    setTouchStart(e.targetTouches[0].clientX);
-  };
-  const onTouchMove = (e: any) => setTouchEnd(e.targetTouches[0].clientX);
-  const onTouchEnd = () => {
-    if (!touchStart || !touchEnd) {
-      return;
-    }
-    const distance = swipeDistance();
-    const isLeftSwipe = distance > minSwipeDistance;
-    if (isLeftSwipe) {
-      onOpenDeleteExercise();
-      setOffset(distance);
-    }
-  };
-  const customOnCloseDeleteExercise = () => {
-    setOffset(0);
-    onCloseDeleteExercise();
-  };
-
-  // Delete Exercise Modal Controls
   const {
-    isOpen: isOpenDeleteExercise,
-    onOpen: onOpenDeleteExercise,
-    onClose: onCloseDeleteExercise,
-  } = useDisclosure();
+    completedSets,
+    customTitle,
+    customWeight,
+    errors,
+    isOpenDeleteExercise,
+    minSwipeDistance,
+    offset,
+    seeDetails,
+    timerIsActive,
+    repsDisplayIsInvalid,
+    repsIsInvalid,
+    setsIsInvalid,
+    timerIsInvalid,
+    titleIsInvalid,
+    weightIsInvalid,
+    weightUnitIsInvalid,
+    completedASet,
+    customOnCloseDeleteExercise,
+    handleTimerIsActive,
+    onDeleteExercise,
+    onOpenDeleteExercise,
+    onTouchEnd,
+    onTouchMove,
+    onTouchStart,
+    removedASet,
+    setCustomTitle,
+    setCustomWeight,
+    setExerciseComment,
+    setOffset,
+    setSeeDetails,
+    setTime,
+    swipeDistance,
+  } = useCreateExerciseForm({
+    exercise,
+    handleExercise,
+    deleteExercise,
+    exerciseIndex,
+    setFormHasErrors,
+  });
 
   return (
     <ExerciseContainer
